qmake, moc and core builds without warnings on win64 (sizeof(int)!=sizeof(size_t)), except 3dparty stuff and warnings in the sdk itself
[git-p4: depot-paths = "//depot/qt/main/": change = 161625]

+++ qmake/generators/makefile.cpp
@@ -1416,7 +1416,7 @@ MakefileGenerator::writeExtraCompilerTargets(QTextStream &t)
                         if(FILE *proc = QT_POPEN(dep_cmd.latin1(), "r")) {
                             QString indeps;
                             while(!feof(proc)) {
-                                int read_in = fread(buff, 1, 255, proc);
+                                int read_in = (int)fread(buff, 1, 255, proc);
                                 if(!read_in)
                                     break;
                                 indeps += QByteArray(buff, read_in);
@@ -1456,7 +1456,7 @@ MakefileGenerator::writeExtraCompilerTargets(QTextStream &t)
                     if(FILE *proc = QT_POPEN(dep_cmd.latin1(), "r")) {
 	                QString indeps;
                         while(!feof(proc)) {
-                            int read_in = fread(buff, 1, 255, proc);
+                            int read_in = (int)fread(buff, 1, 255, proc);
                             if(!read_in)
                                 break;
                             indeps += QByteArray(buff, read_in);
